refactor to use standard Linux config path
